+++ CHANGELOG.md
@@ -6,11 +6,14 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Example for Issue #1716 showing poor tick spacing on DateTimeAxis with interval types of Weeks or Years
 - Example for label placement on BarSeries with non-zero BaseValue (#1726)
+- ExtrapolationLineSeries as described in #1740 to allow to display pre-defined intervals of a line series with a different style than the rest.
+- Added DataRange, unit tests and examples as support for ExtrapolationLineSeries.
 - Add control over how far from the series the tracker fires (#1736)
 - Add option to check distance for result between data points (#1736)
 - Legend.AllowUseFullExtent property to control whether legends should be able to use the full extent of the plot (#1743)
 
 ### Changed
+- Updated Series.cd with ExtrapolationLineSeries and removed classes that do not exist anymore
 
 ### Removed